Remove old DropletConfiguration
